factors out video_data_generation
