+++ src/generation/mod.rs
@@ -1,2 +1,3 @@
 pub mod parsed_to_render;
+pub mod video_data_generation;
 mod test;
