+++ src/generation/parsed_to_render.rs
@@ -9,11 +9,13 @@ use crate::write::{write_composition_to_json, write_composition_to_wav};
 use num_rational::Rational64;
 use pbr::ProgressBar;
 use rayon::prelude::*;
-use serde::{Deserialize, Serialize};
 use serde_json::to_string;
 use socool_ast::ast::OpOrNfTable;
 use socool_ast::operations::{NormalForm, Normalize as NormalizeOp, PointOp};
 use std::sync::{Arc, Mutex};
+use crate::generation::video_data_generation::{
+    vec_timed_op_to_vec_op4d, composition_to_vec_timed_op,
+};
 
 pub fn r_to_f64(r: Rational64) -> f64 {
     *r.numer() as f64 / *r.denom() as f64
@@ -41,114 +43,6 @@ pub fn to_wav(composition: StereoWaveform, filename: String) {
     banner("Printing".to_string(), filename.clone());
     write_composition_to_wav(composition, &filename);
     printed("WAV".to_string());
-}
-
-#[derive(Debug, Clone, Eq, Ord, PartialEq, PartialOrd, Serialize, Deserialize)]
-pub enum EventType {
-    On,
-    Off,
-}
-
-#[derive(Debug, Clone, Eq, Ord, PartialEq, PartialOrd)]
-pub struct TimedOp {
-    t: Rational64,
-    event_type: EventType,
-    voice: usize,
-    event: usize,
-    fm: Rational64,
-    fa: Rational64,
-    pm: Rational64,
-    pa: Rational64,
-    g: Rational64,
-    l: Rational64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct Op4D {
-    t: f64,
-    event_type: EventType,
-    voice: usize,
-    event: usize,
-    x: f64,
-    y: f64,
-    z: f64,
-    l: f64,
-}
-
-impl TimedOp {
-    fn to_op_4d(&self, basis: &Basis) -> Op4D {
-        Op4D {
-            l: r_to_f64(self.l) * basis.l,
-            t: r_to_f64(self.t) * basis.l,
-            x: ((basis.p + r_to_f64(self.pa)) * r_to_f64(self.pm)),
-            y: (basis.f * r_to_f64(self.fm)) + r_to_f64(self.fa),
-            z: basis.g * r_to_f64(self.g),
-            voice: self.voice,
-            event: self.event,
-            event_type: self.event_type.clone(),
-        }
-    }
-}
-
-fn point_op_to_timed_op(
-    point_op: &PointOp,
-    time: &mut Rational64,
-    voice: usize,
-    event: usize,
-) -> (TimedOp, TimedOp) {
-    let on = TimedOp {
-        fm: point_op.fm,
-        fa: point_op.fa,
-        pm: point_op.pm,
-        pa: point_op.pa,
-        g: point_op.g,
-        l: point_op.l,
-        t: time.clone(),
-        event_type: EventType::On,
-        voice,
-        event,
-    };
-
-    *time += point_op.l;
-
-    let off = TimedOp {
-        t: time.clone(),
-        event_type: EventType::Off,
-        ..on
-    };
-
-    (on, off)
-}
-
-fn vec_timed_op_to_vec_op4d(timed_ops: Vec<TimedOp>, basis: &Basis) -> Vec<Op4D> {
-    timed_ops.iter().map(|t_op| t_op.to_op_4d(&basis)).collect()
-}
-
-fn composition_to_vec_timed_op(composition: &NormalForm, table: &OpOrNfTable) -> Vec<TimedOp> {
-    let mut normal_form = NormalForm::init();
-
-    println!("Generating Composition \n");
-    composition.apply_to_normal_form(&mut normal_form, table);
-
-    let mut result: Vec<TimedOp> = normal_form
-        .operations
-        .iter()
-        .enumerate()
-        .flat_map(|(voice, vec_point_op)| {
-            let mut time = Rational64::new(0, 1);
-            let mut result = vec![];
-            vec_point_op.iter().enumerate().for_each(|(event, p_op)| {
-                let (on, off) = point_op_to_timed_op(p_op, &mut time, voice, event);
-                result.push(on);
-                result.push(off);
-            });
-            result
-        })
-        .collect();
-
-    result.sort_unstable_by_key(|a| a.t);
-
-    result
 }
 
 pub fn to_json(basis: &Basis, composition: &NormalForm, table: &OpOrNfTable, filename: String) {
